arch/powerpc: cleanup systemsim(mambo) support and add special idle routine
The original patchset, here:
https://lists.ozlabs.org/pipermail/linuxppc64-dev/attachments/20060113/eb4deb3e/attachment.txt
added a SYSTEMSIM platform. Let's do that same thing so all the config
options can be disabled easily. Make it depend on PPC_POWERNV for when
it is run with skiboot and PPC_PSERIES for when it is run with the
simulated RTAS layer.

The current public releases of Mambo (P8 and P9) are officially called
the "IBM POWER{8,9} Functional Simulator". Update the help text and
other names to reflect this.

Update bogus net so that we only call irq_create_mapping() if mambo
gives us a non-zero irq back from the bogus net probe call (if we get
zero we are operating in polled mode).

And finally, add a special idle routine for when we are running on the
simulator. This is based on the patch posted to linuxppc64-dev by Arnd
Bergmann. ([patch 3/8] powerpc: cpu_idle implementation for systemsim
https://lists.ozlabs.org/pipermail/linuxppc64-dev/2006-January/007621.html)

I've updated it so that systemsim_halt() can be used as a
ppc_md.power_save() routine.

The original patch hooked this in to setup_arch(), but we only need
it if we're running with the emulated RTAS layer (not when we're
running with skiboot), so let's hook it in to pSeries_setup_arch().

Signed-off-by: Mark Nelson <nelsondm@us.ibm.com>